Split Flesh.Language.Parser.Class module from Error
