+++ hspec-src/Flesh/Language/Parser/TestUtil.hs
@@ -28,8 +28,8 @@ import qualified Data.Map.Strict as M
 import qualified Data.Text as T
 import qualified Flesh.Language.Alias as Alias
 import Flesh.Language.Parser.Char
+import Flesh.Language.Parser.Class
 import Flesh.Language.Parser.Error
---import Flesh.Language.Parser.Input
 import Flesh.Source.Position
 import Test.Hspec
 import Test.Hspec.QuickCheck
