+++ src/Flesh/Language/Parser/Alias.hs
@@ -42,6 +42,7 @@ import Control.Monad.Trans.Maybe
 import qualified Data.Map.Strict as M
 import qualified Data.Text as T
 import Flesh.Language.Alias
+import Flesh.Language.Parser.Class
 import Flesh.Language.Parser.Error
 import Flesh.Language.Parser.Input
 import Flesh.Source.Position
