+++ src/Flesh/Language/Parser/Lex.hs
@@ -35,6 +35,7 @@ import Data.Char
 import qualified Data.List.NonEmpty as NE
 import Flesh.Source.Position
 import Flesh.Language.Parser.Char
+import Flesh.Language.Parser.Class
 import Flesh.Language.Parser.Error
 import Flesh.Language.Parser.Input
 import Numeric.Natural
